Rename and move some stuff
