+++ apps/server/src/modules/user-management/uc/dto/class-for-user-list.dto.ts
@@ -1,11 +1,11 @@
 import { EntityId } from '@shared/domain/types';
 
-export class ClassDto {
+export class ClassForUserListDto {
 	public id: EntityId;
 
 	public name: string;
 
-	constructor(props: ClassDto) {
+	constructor(props: ClassForUserListDto) {
 		this.id = props.id;
 		this.name = props.name;
 	}
