+++ apps/server/src/modules/user-management/uc/mapper/class.dto.mapper.ts
@@ -1,11 +1,11 @@
 import { Class } from '../../domain/class';
-import { ClassDto } from '../dto/class.dto';
+import { ClassForUserListDto } from '../dto/class-for-user-list.dto';
 
 export class ClassDtoMapper {
-	public static mapToDto(c: Class): ClassDto {
+	public static mapToDto(c: Class): ClassForUserListDto {
 		const classProps = c.getProps();
 
-		const dto = new ClassDto({
+		const dto = new ClassForUserListDto({
 			id: classProps.id,
 			name: classProps.name,
 		});
@@ -13,7 +13,7 @@ export class ClassDtoMapper {
 		return dto;
 	}
 
-	public static mapToDtos(classes: Class[]): ClassDto[] {
+	public static mapToDtos(classes: Class[]): ClassForUserListDto[] {
 		const dtos = classes.map((c) => ClassDtoMapper.mapToDto(c));
 
 		return dtos;
