+++ apps/server/src/modules/user-management/uc/type/sortable-fields.ts
@@ -1,0 +1,9 @@
+export enum SortableFields {
+	firstName = 'firstName',
+	lastName = 'lastName',
+	birthday = 'birthday',
+	email = 'email',
+	class = 'class',
+	registration = 'registration',
+	createdAt = 'createdAt',
+}
